requirements.txt
fastapi>=0.120.0
uvicorn[standard]>=0.38.0
sse-starlette>=3.0.0
openai>=1.0.0
pydantic>=2.12.0
python-dotenv>=1.0.0

